# swagger-example.json
{
    "basePath": "/example/api/v1",
    "paths": {
...
        "/types": {
            "post": {
                "produces": ["application/json"],
                "description": "Creates a new type",
                "operationId": "posttypes",
                "responses": {
                    "201": {
                        "schema": {"$ref": "#/definitions/Type"},
                        "description": "Type entity created successfully"
                    },
                    "400": {"description": "An error occurred while validation"}
                },
                "parameters": [{
                    "schema": {"$ref": "#/definitions/Type"},
                    "in": "body",
                    "name": "type",
                    "description": "Type entity ",
                    "required": true
                }],
                "consumes": ["application/json"],
                "tags": ["types"]
            },
            "get": {
                "produces": ["application/json"],
                "description": "",
                "operationId": "gettypes",
                "responses": {
                    "200": {
                        "schema": {
                            "type": "array",
                            "items": {"$ref": "#/definitions/Type"}
                        },
                        "description": "Type entities found"
                    },
                    "500": {"description": "An error occured while request"}
                },
                "parameters": [
                    {
                        "in": "query",
                        "name": "name",
                        "description": "name of type or null ",
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "createdAt",
                        "description": "creation date range or null ",
                        "type": "string"
                    }
                ],
                "tags": ["types"]
            },
            "delete": {
                "description": "This method is not allowed. When requesting the HTTP status 405 will return",
                "operationId": "deletetypes",
                "responses": {"405": {"description": "Method not allowed"}},
                "tags": ["types"]
            },
            "put": {
                "description": "This method is not allowed. When requesting the HTTP status 405 will return",
                "operationId": "puttypes",
                "responses": {"405": {"description": "Method not allowed"}},
                "tags": ["types"]
            }
        }
    },
    "host": "localhost:8080",
    "definitions": {"Type": {
        "type": "object",
        "properties": {
            "createdAt": {
                "format": "date-time",
                "type": "string"
            },
            "codes": {
                "type": "array",
                "items": {
                    "format": "long",
                    "type": "number"
                }
            },
            "byteString": {
                "format": "byte",
                "type": "string"
            },
            "bool": {"type": "boolean"},
            "name": {"type": "string"},
            "id": {
                "format": "long",
                "type": "number"
            },
            "something": {
                "type": "array",
                "items": {"type": "string"}
            }
        }
    }},
    "swagger": "2.0",
    "info": {
        "description": "Not existing API for demonstrating the javadoc2swagger-maven-plugin",
        "title": "Example API",
        "version": "0.0.1-SNAPSHOT"
    },
...
}